cleaning up example a bit. Still not done
