--- example/README.md
@@ -1,1 +1,0 @@
-Open this folder in a terminal and run `node example-server` to start the example server.
