+++ example/public/index.html
@@ -2,30 +2,94 @@
 <html>
 <head lang="en">
   <meta charset="UTF-8">
-  <title></title>
+  <title>Feathers Authentication</title>
 </head>
 <body>
-  <h1>feathers-passport</h1>
-  <a href="/api/todos/dishes">REST endpoint /todos/rest</a><br/><br/>
-  <a href="/sockets.html">Test sockets</a>
-  <pre id="log"></pre>
-  
-  <h3>Login</h3>
-  <p>feathers / test</p>
-  <form action="/api/login" method="post">
-    <div>
-      <label>Username:</label>
-      <input type="text" name="username"/>
-    </div>
-    <div>
-      <label>Password:</label>
-      <input type="password" name="password"/>
-    </div>
-    <div>
-      <input type="submit" value="Log In"/>
-    </div>
-  </form>
-  <a href="/logout.html">Logout</a>
+  <h1>Feathers Local Authentication Examples</h1>
 
+  <section>
+    <h3>Email/Password authentication via REST</h3>
+    <p>Username: <code>admin@feathersjs.com</code></p>
+    <p>Password: <code>admin</code></p>
+
+    <form action="/auth/local" method="post">
+      <div>
+        <label>Email:</label>
+        <input type="text" name="email" value="admin@feathersjs.com"/>
+      </div>
+      <div>
+        <label>Password:</label>
+        <input type="password" name="password" value="admin"/>
+      </div>
+
+      <button id="email-password-rest-button">Login</button>
+    </form>
+  </section>
+
+  <section>
+    <h3>Token based authentication via REST</h3>
+
+    <form action="/auth/token" method="post">
+      <div>
+        <label>Token:</label>
+        <input type="text" name="token" value=""/>
+      </div>
+      
+      <button id="token-rest-button">Login</button>
+    </form>
+  </section>
+
+  <section>
+    <h3>Basic email/password authentication via socket</h3>
+    <form action="/auth/local" method="post">
+      <div>
+        <label>Email:</label>
+        <input type="text" name="email" value="admin@feathersjs.com"/>
+      </div>
+      <div>
+        <label>Password:</label>
+        <input type="password" name="password" value="admin"/>
+      </div>
+
+      <button id="email-password-socket-button">Login</button>
+    </form>
+  </section>
+
+  <section>
+    <h3>Token based authentication via socket</h3>
+    <form action="/auth/token" method="post">
+      <div>
+        <label>Token:</label>
+        <input type="text" name="token" value=""/>
+      </div>
+
+      <button id="token-socket-button">Login</button>
+    </form>
+  </section>
+  
+  <pre id="log"></pre>
+
+  <script src="socket.io/socket.io.js" type="text/javascript"></script>
+  <script type="text/javascript">
+    // Normally, you would get this from /login or if you stored it in localStorage, from there.
+    var token = 'eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJlbWFpbCI6ImFkbWluQGZlYXRoZXJzanMuY29tIiwiaWQiOjAsImlhdCI6MTQ1NDA5MzI0NH0.6aDm7KF2MWnlNNBiQfdCXRczWN7LOhBWLP_NIj0xy6o';
+    var socket = io('', {
+      query: 'token=' + token,
+      transport: ['websockets']
+    });
+    var el = document.getElementById('log');
+
+    socket.on('error', function() {
+      console.error(arguments);
+    });
+    socket.on('connect', function() {
+      socket.emit('messages::find', {}, function(error, messages) {
+        messages.forEach(function(message){
+          el.innerHTML += message.text + '<br/>';
+        });
+      });
+    });
+    
+  </script>
 </body>
 </html>
