--- example/public/logout.html
@@ -1,10 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head lang="en">
-  <meta charset="UTF-8">
-  <title>Logout</title>
-</head>
-<body>
-  There's no need to logout because the server isn't stateful (no sessions).  Just destroy the auth token and you're done.
-</body>
-</html>
